+++ client/search-ui/src/index.ts
@@ -1,6 +1,5 @@
 export * from './components'
 export * from './documentation/ModalVideo'
-export * from './input/CodeMirrorQueryInputNG'
 export * from './input/CodeMirrorQueryInput'
 export * from './input/codemirror'
 export * from './input/LazyMonacoQueryInput'
@@ -19,4 +18,3 @@ export * from './results/sidebar/SearchReference'
 export * from './results/sidebar/SearchTypeLink'
 export * from './results/StreamingSearchResultsList'
 export * from './util'
-export { Source, Group, Option } from './input/codemirror/suggestions'
