+++ client/search-ui/src/input/experimental/HighlightedSearchQuery.svelte
