+++ client/search-ui/src/input/experimental/Icon.svelte
