+++ client/search-ui/src/input/experimental/README.md
@@ -1,0 +1,6 @@
+This folder contains a new (experimental) implementation of the search query
+input.
+
+IT SHOULD NOT BE USED!
+
+If you have questions about it reach out the search product team.
