+++ client/search-ui/src/input/experimental/SearchQueryOption.svelte
