+++ client/search-ui/src/input/experimental/index.ts
@@ -1,0 +1,4 @@
+export { LazyCodeMirrorQueryInput } from './LazyCodeMirrorQueryInput'
+export { Group, Option, EntryOf, Source, getEditorConfig } from './suggestions'
+export { default as SearchQueryOption } from './SearchQueryOption.svelte'
+export { default as FilterOption } from './FilterSuggestion.svelte'
